mcp-severa/requirements.txt
google-adk
fastapi 
uvicorn
fastmcp
python-dotenv
